@@ -34,47 +34,47 @@ class BuiltInAdapterTest extends TestCase
     public static function throwawayDomains()
     {
         // List of some of the built-in throwaway domains
-        return array(
-            array('throwawayemailaddress.com'),
-            array('tmail.com'),
-            array('tmailinator.com'),
-            array('tradermail.info'),
-            array('trash-amil.com'),
-            array('trash-mail.com'),
-            array('trash-mail.de'),
-            array('trash2009.com'),
-            array('trashemail.de'),
-            array('trashmail.at'),
-            array('trashmail.com'),
-            array('trashmail.net'),
-            array('trashmail.ws'),
-            array('trashmailer.com'),
-            array('trashymail.com'),
-            array('trashymail.net'),
-            array('trillianpro.com'),
-            array('turual.com'),
-            array('tyldd.com'),
-            array('uggsrock.com'),
-            array('upliftnow.com'),
-            array('uplipht.com'),
-            array('veryrealemail.com'),
-            array('viditag.com'),
-            array('viewcastmedia.com'),
-            array('viewcastmedia.net'),
-            array('viewcastmedia.org'),
-            array('webm4il.info'),
-            array('wegwerfemail.de'),
-            array('wetrainbayarea.com'),
-            array('wetrainbayarea.org'),
-            array('wh4f.org'),
-            array('whyspam.me'),
-            array('willselfdestruct.com'),
-            array('wuzup.net'),
-            array('wuzupmail.net'),
-            array('xagloo.co'),
-            array('xemaps.com'),
-            array('xmail.com'),
-            array('yopmail.com'),
-        );
+        return [
+            ['throwawayemailaddress.com'],
+            ['tmail.com'],
+            ['tmailinator.com'],
+            ['tradermail.info'],
+            ['trash-amil.com'],
+            ['trash-mail.com'],
+            ['trash-mail.de'],
+            ['trash2009.com'],
+            ['trashemail.de'],
+            ['trashmail.at'],
+            ['trashmail.com'],
+            ['trashmail.net'],
+            ['trashmail.ws'],
+            ['trashmailer.com'],
+            ['trashymail.com'],
+            ['trashymail.net'],
+            ['trillianpro.com'],
+            ['turual.com'],
+            ['tyldd.com'],
+            ['uggsrock.com'],
+            ['upliftnow.com'],
+            ['uplipht.com'],
+            ['veryrealemail.com'],
+            ['viditag.com'],
+            ['viewcastmedia.com'],
+            ['viewcastmedia.net'],
+            ['viewcastmedia.org'],
+            ['webm4il.info'],
+            ['wegwerfemail.de'],
+            ['wetrainbayarea.com'],
+            ['wetrainbayarea.org'],
+            ['wh4f.org'],
+            ['whyspam.me'],
+            ['willselfdestruct.com'],
+            ['wuzup.net'],
+            ['wuzupmail.net'],
+            ['xagloo.co'],
+            ['xemaps.com'],
+            ['xmail.com'],
+            ['yopmail.com'],
+        ];
     }
 }
